requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.5.1
numpy==1.26.4
reportlab==4.2.4
PyJWT==2.8.0
gunicorn==23.0.0
Werkzeug==3.0.3
tensorflow==2.16.1